backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.25.2
anthropic==0.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0
pytube==15.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
gunicorn==21.2.0
